Moved all Weapons into one package
instead of one package per class, it's one package for all weapons with 4 sub-packages (one for each class). This fits in better with the enemies pacakge, guiPieces package, and the images folder.
