+++ src/drivers/GuiController.java
@@ -23,29 +23,29 @@ import buildComparators.CompareAccuracyGraphs;
 import buildComparators.CompareMetrics;
 import dataGenerator.DatabaseConstants;
 import dataGenerator.MetricsCalculator;
-import drillerWeapons.CryoCannon;
-import drillerWeapons.EPC_ChargeShot;
-import drillerWeapons.EPC_RegularShot;
-import drillerWeapons.Flamethrower;
-import drillerWeapons.Subata;
-import engineerWeapons.BreachCutter;
-import engineerWeapons.BreachCutter_Projectile;
-import engineerWeapons.GrenadeLauncher;
-import engineerWeapons.SMG;
-import engineerWeapons.Shotgun;
 import guiPieces.HoverText;
 import guiPieces.View;
-import gunnerWeapons.Autocannon;
-import gunnerWeapons.BurstPistol;
-import gunnerWeapons.Minigun;
-import gunnerWeapons.Revolver;
 import modelPieces.EnemyInformation;
-import modelPieces.Weapon;
-import scoutWeapons.Boomstick;
-import scoutWeapons.Classic_FocusShot;
-import scoutWeapons.Classic_Hipfire;
-import scoutWeapons.AssaultRifle;
-import scoutWeapons.Zhukov;
+import weapons.Weapon;
+import weapons.driller.CryoCannon;
+import weapons.driller.EPC_ChargeShot;
+import weapons.driller.EPC_RegularShot;
+import weapons.driller.Flamethrower;
+import weapons.driller.Subata;
+import weapons.engineer.BreachCutter;
+import weapons.engineer.BreachCutter_Projectile;
+import weapons.engineer.GrenadeLauncher;
+import weapons.engineer.SMG;
+import weapons.engineer.Shotgun;
+import weapons.gunner.Autocannon;
+import weapons.gunner.BurstPistol;
+import weapons.gunner.Minigun;
+import weapons.gunner.Revolver;
+import weapons.scout.AssaultRifle;
+import weapons.scout.Boomstick;
+import weapons.scout.Classic_FocusShot;
+import weapons.scout.Classic_Hipfire;
+import weapons.scout.Zhukov;
 
 /*
 	Benchmarks:
