+++ src/weapons/gunner/Autocannon.java
@@ -1,4 +1,4 @@
-package gunnerWeapons;
+package weapons.gunner;
 
 import java.util.ArrayList;
 
@@ -14,9 +14,9 @@ import modelPieces.Mod;
 import modelPieces.Overclock;
 import modelPieces.StatsRow;
 import modelPieces.UtilityInformation;
-import modelPieces.Weapon;
 import utilities.ConditionalArrayList;
 import utilities.MathUtils;
+import weapons.Weapon;
 
 public class Autocannon extends Weapon {
