+++ src/weapons/gunner/Minigun.java
@@ -1,4 +1,4 @@
-package gunnerWeapons;
+package weapons.gunner;
 
 import java.util.ArrayList;
 
@@ -14,10 +14,10 @@ import modelPieces.Mod;
 import modelPieces.Overclock;
 import modelPieces.StatsRow;
 import modelPieces.UtilityInformation;
-import modelPieces.Weapon;
 import spreadCurves.MinigunCurve;
 import utilities.ConditionalArrayList;
 import utilities.MathUtils;
+import weapons.Weapon;
 
 public class Minigun extends Weapon {
