+++ src/weapons/gunner/Revolver.java
@@ -1,4 +1,4 @@
-package gunnerWeapons;
+package weapons.gunner;
 
 import java.util.ArrayList;
 
@@ -13,10 +13,10 @@ import modelPieces.Mod;
 import modelPieces.Overclock;
 import modelPieces.StatsRow;
 import modelPieces.UtilityInformation;
-import modelPieces.Weapon;
 import spreadCurves.RevolverCurve;
 import utilities.ConditionalArrayList;
 import utilities.MathUtils;
+import weapons.Weapon;
 
 public class Revolver extends Weapon {
