+++ src/weapons/scout/AssaultRifle.java
@@ -1,4 +1,4 @@
-package scoutWeapons;
+package weapons.scout;
 
 import java.util.ArrayList;
 
@@ -14,10 +14,10 @@ import modelPieces.Mod;
 import modelPieces.Overclock;
 import modelPieces.StatsRow;
 import modelPieces.UtilityInformation;
-import modelPieces.Weapon;
 import spreadCurves.AssaultRifleCurve;
 import utilities.ConditionalArrayList;
 import utilities.MathUtils;
+import weapons.Weapon;
 
 public class AssaultRifle extends Weapon {
