+++ src/weapons/scout/Boomstick.java
@@ -1,4 +1,4 @@
-package scoutWeapons;
+package weapons.scout;
 
 import java.util.ArrayList;
 
@@ -13,9 +13,9 @@ import modelPieces.Mod;
 import modelPieces.Overclock;
 import modelPieces.StatsRow;
 import modelPieces.UtilityInformation;
-import modelPieces.Weapon;
 import utilities.ConditionalArrayList;
 import utilities.MathUtils;
+import weapons.Weapon;
 
 /*
 	Extracted via UUU:
